Database, Routes, cars by categories, Views

+++ Shop/Data/Models/Car.cs
@@ -12,7 +12,7 @@ namespace Shop.Data.Models
 
         public string img { get; set; }
 
-        public ushort price { get; set; }
+        public int price { get; set; }
         
         public bool isFavourite { get; set; }
